Adding Profile route linked to by profile icon
